# stop using deepcopy in compound spliter test to speed up